refactor(snapshot): keep backend capability metadata internal
